Move `Application` inheritors to `Application.kt` file.
